refactor: move tenant benchmark tests to a dedicated package for better consistency with other bounded context
