+++ tests/bench/tenant/namespace_repo_test.go
@@ -1,7 +1,7 @@
 //go:build !unit_test
 // +build !unit_test
 
-package bench
+package tenant
 
 import (
 	"context"
